# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PedagangPulsa API
  description: |
    PedagangPulsa Member Mobile API v1.0

    ## Authentication

    The API uses JWT (JSON Web Token) based authentication with a 2-step verification process:

    1. **Login** → Get Access Token (15min) + Refresh Token (7days)
    2. **PIN Verify** → Get PIN Session Token (5min) for sensitive operations

    ## Idempotency

    Transaction and Transfer endpoints require `X-Reference-Id` header to prevent duplicate requests.

  version: 1.0.0
  contact:
    name: PedagangPulsa API Support
    email: api@pedagangpulsa.com
    url: https://pedagangpulsa.com
  license:
    name: Proprietary
servers:
  - url: https://api.pedagangpulsa.com/api/v1
    description: Production server
  - url: https://sandbox-api.pedagangpulsa.com/api/v1
    description: Sandbox server
...
paths:
...
  /balance:
    get:
      tags:
        - Balance
      summary: Get current balance
      description: Get user's current balance information
      operationId: getBalance
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Balance retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Balance'
  /balance/history:
    get:
      tags:
        - Balance
      summary: Get balance history
      description: Get user's balance mutation history with pagination
      operationId: getBalanceHistory
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Balance history retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BalanceHistoryResponse'
...
  /notifications/{id}/read:
    post:
      tags:
        - Notifications
      summary: Mark notification as read
      operationId: markNotificationRead
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /notifications/read-all:
    post:
      tags:
        - Notifications
      summary: Mark all notifications as read
      operationId: markAllNotificationsRead
      security:
        - BearerAuth: []
      responses:
        '200':
          description: All notifications marked as read
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          marked_count:
                            type: integer
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /auth/login
...
  schemas:
    SuccessResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total_items:
          type: integer
          example: 150
        total_pages:
          type: integer
          example: 8

    # Auth schemas
...
    Balance:
      type: object
      properties:
        active_balance:
          type: integer
          example: 150000
        held_balance:
          type: integer
          example: 5600
        total_balance:
          type: integer
          example: 155600
        currency:
          type: string
          example: IDR
    BalanceHistoryResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              type:
                type: string
                enum: [topup, transaction, refund, adjustment, referral, transfer_in, transfer_out]
              amount:
                type: integer
              balance_after:
                type: integer
              description:
                type: string
              ref_type:
                type: string
              ref_id:
                type: string
                format: uuid
              created_at:
                type: string
                format: date-time
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Topup schemas
...